Properly handle ProtectionExtensions as singleton
Moved set of available key systems into setup() method.

ProtectionExtensions.autoSelectKeySystem() now takes
ProetectionModel as a parameter.  This prevents our singleton
from depending on a non-singleton as a member variable.

ProtectionExtensions.init() (which sets the protectionData in
each key sysytem) is now called in StreamController.  Since
StreamController is also a singleton, this ensures that init()
is called only once.