Fix multiple LicenseRequestComplete subscriptions
When playing content, multiple 'needkey' events will be delivered
over the duration of the media.  In my handler, this is where we
detect the key system to be used and then subscribe to
LicenseRequestComplete events.  This subscription only needs to
be done the first time.

Prevent key system changes within a single stream.  It is
illegal to change key systems within a Period.

Also, teardown the protection system of the stream before tearing
down the MediaSource.  Chrome complains if you try to cancel
KeySessions if there is no media present.