Fix 01b multi-session bug
Queue up pending sessions until we receive the associated key
message

When multiple keys are requested initially, the createKeySession
calls in before the first keymessage is generated.  This means that
we can't detect whether or not multiple key sessions are supported
until after multiple key sessions have already been created.  This
is something that we can't avoid with this early EME API version.

The outcome is that if you try to play encrypted content that uses
more than one key on a user agent that implements the 01b API and
DOES NOT support sessionIDs, you will get unpredictable results.

At this point, I do not know of any browsers that will cause this
error, so I just log an error message