Support initData in MPD for ClearKey and Widevine
Added a standard way to support initData in the MPD for both
Widevine and ClearKey.  A Widevine representative has confirmed
that this is the standard way to format their ContentProtection
element.  For ClearKey, I have chosen to use the same format,
since it is very simply and works well.

Updated the sources.json to add new test content for common encryption
with and without init data present in the MPD.

Removed PlayReady/ClearKey content from the sources.json.  IE11 does
not support PSSH version 1 (which ClearKey requires).  Just the presence
of the ClearKey PSSH in the file causes an error, so we won't be able
to test this combination of DRMs until the next version of IE (Spartan)
is released.