Added a feature to make selecting a starting bit rate a little more intelligent.
During playback when a quality level changes it is recorded into local storage now. 
Upon starting a stream this information is retrieved and if it not expired (< 1 hour old) then it is used to pick the starting quality for playback. If it is not present or expired we default back to the bit rate nearest 1000kpbs