fixed issue that would override explicit setting of initial bit rate from player level.  Now it detects if there is an explicitly set initial bit rate before overriding with local storage value if it exists and finally defaulting to the 1000/100 kbps value if none of the above exists.
Also removed resetting bitrateDict in abrcontroller. We should let this setting persist over multiple sources. If we empty the object each time we attach source in media player then there is no way to set initial bit rate for the second media source and so on...

Also moved where the default value is set (from abrcontroller to new checkInitialBitrate in mediaplayer) so I have something to use to detect if player set initial bit rate.