Simplify and improve manifest handling
ManifestLoader is no longer a singleton so it can be used outside
the player simulataneously without having to worry about race conditions
around MANIFEST_LOADED events

ManifestUpdater is now the single point of access for manifest update
events.  Classes inside the player should no longer register for
MANIFEST_LOADED events -- only for MANIFEST_UPDATED events.
ManifestUpdater is exclusively responsible for updating the
ManifestModel singleton object that the rest of the player relies upon
for the latest manifest object

StreamController is now responsible for instantiating the ManifestLoader
instance for the player and (optionally) loading a manifest from a URL

It is now possible to attach an already loaded/parsed manifest to the
MediaPlayer

DashAdapter is a utility that can be used outside the player to extract
interesting information from the manifest.  However, it was previously
dependent on the singleton ManifestModel object which holds the single
manifest that the player is currently using.  All DashAdapter methods
(and the code that calls them) have been updated to require a manifest
object as an argument when needed.