ManifestLoader: Initialize baseUrl to "" instead of null.
Initialize baseUrl to "" instead of null so that if the Manifest
URL is relative without a "/", then we return an empty baseUrl
rather than null. Without this change, if a manifest url is
relative without a "/", then the media URL will be resolved as
null + media_url. Where as the correct media URL will be
"" + media_url.