Cleanup manifest updater
start() was not called and stop() was only called from streamController.reset(), so
removed both methods and created a reset().

By default, updating is stopped and is only started when playback begins.  Updating
is stopped only when playback is paused.  However, this doees not prevent us from
receiving updates when a manifest is loaded or "set" via setManfiest().