Fix fallback-to-needkey mechanism
When initData is found in the MPD, we need to check for failures in
both "KEY_SYSTEM_ACCESS" and "KEY_SYSTEM_SELECTED" events.  Previously,
we were only checking "KEY_SYSTEM_SELECTED" errors.  Also moved the code
that falls back to needkey events into a function since it needs to be
called from several places.

Also, used the "oneShot" mechanism for event subscription.  When trying to
select a key system from initData, we only try once, then we fall-back to
needkey, so this saves us call to "unsubscribe".