Move LICENSE_REQUEST_COMPLETE event notification
Since ProtectionExtensions is a singleton, all of the
ProtectionControllers/Models that might exist can't use it as their
source for license request events.  Otherwise, each controller won't
know which event is for them.  This change moves the event subscription
and notification into the ProtectionController and leaves just the
process of retrieving the license in ProtectionExt.