Fix for bug #556
When an attempt to select a key system fails, we need to reset the
ProtectionController.keySystem property to undefined.  This returns us
fully to the "unselected" state.  When keySystem is null, this is our
"in progress" state and it just queues up subsequent requests for
key system selection.

In the case of the test content that exhibited the bug, we first tried
to select a key system from the ContentProtection elements in the MPD.
When this failed, we remained in the "in progress" state and all future
attempts to select a key system were queued up instead of being properly
handled.