Fix inverted boolean logic
In my commit for the new protection event model, I renamed the
boolean parameter for ProtectionController.selectKeySystem to something
that I felt made more sense.  This renaming reversed the boolean logic
for that parameter.  I forgot to change the places where this function
is called to invert the boolean parameter.