Fix sourcemaps for dash.js code. #977
This produces sourcemap files that are valid for dash.js code.

When running uglify with the beautify parameter set to true, some sourcemaps are corrupted.
For an example you can enable maps for .debug and add a breakpoint in Stream.js setup function
which results in a breakpoint in the BufferController onCurrentTrackChanged function.