Apply requestModifierExt to the headers on manifest load.
When loading fragments, the requestModifierExt is applied to both the
URL and the headers. This allows doing things like setting
`withCredentials` on each fragment request. Apply the same handling to
the manifest load in case the manifest load and fragment loads need to
be routed to the same server in a load balanced situation.