Allow providing protData without requiring serverURL
Currently if the end user provides any protection data for the Key System
our library no longer retrieves the license server URL from the initData.

The above combines with the fact that the ProtectionKeyController skips Key
Systems without protection data if protection data is provided for any
key system in a bad way.

The end result is that if you want to provide some
data for one key system you can't get the serverUrl from the initData for
a different key system. This ensures that even if protection data was passed,
if no serverURL or laURL is provided we try to get the url from the initData.