Fix an exception on eventbus :
EventBus.trigger set type property on payload. When onInitFragmentLoaded is called, chunk is saved in initCache and appended to buffer and an event is triggered (modifying chunk by adding type propery).
If switchInitData is called, then chunk is extracted from initCache. Extracted chunk is appended to buffer and event is triggered -> exception because extracted chunk has a type property.

This commit fix the pb, because chunk is not modified by eventBus