Fix Mss using stream https://djkxw9frc3oss.cloudfront.net/live/Sintel_0500.isml/Manifest
It has to be noticed that using this stream, some fragments are quite long (10s) and sometimes getFragmentRequestForTime doesn't return the right fragment correpsonding to time, but the previous one, because of threshold in FragmentModel:getRequestForTime. That's why there is a loop in nextFragmentRuel, to get the first next not loaded fragment.

Another point is buffer parameters are :
            "liveDelay": 18,
            "stableBufferTime": 16,
            "bufferTimeAtTopQuality": 16,
            "bufferTimeAtTopQualityLongForm": 16