Always add keySystem if supported.
Switch ClearKey module to use correct uuid according to the latest
IOP documentation ('e2719d58-a985-b3c9-781a-b030af78d30e').

In addition, always add the key system to the list of supported
if we match based on the current list of supported key systems.
The requirement of always having a pssh box is not a requirement
for every CDM and if the initData is non-conforming to a particular
key system, creating the session will fail later. We don't need
the clearkey specific condition when determining if a system is
supported -- it should be enough that it's in the list.