fix for Playready Key message format parsing.
Some Playready 2.5 CDMs output the challenge part directly; 'PlayreadyKeyMessage' tag is not included in the key message.

JS player should check for 'PlayreadyKeyMessage' and do the base64 decoding if the tag is present. If it's not found, message should be
passed to the license server.

Also, the content-type header should use utf-8 or utf-16; utf8 & utf16 are not valid encoding types.

Signed-off-by: Gurdal Oruklu <gurdal_oruklu@comcast.com>