Fix: exclude text buffer level if text is disabled
This fix ensure the correct minimum buffer level is reported when text
is disabled. Previously a buffer level of 0 is reported at all times if
a text track is present in the manifest but text is disabled in the
player. This fix checks whether text is enabled and only includes text
buffer metrics in buffer level reporting if it is.