Added subtitle event handling for manual rendering the subtitles (#4360)
* Added subtitle event handling for manual rendering the subtitles

This adds two new media player events: `CUE_ENTER` and `CUE_EXIT`.
With these events you can render the subtitles without relying on
the browser. This makes it possible to fully customize the UI for
the subtitles. To enable the events, set the `fireCueEvents` setting
to true.

For details check out the example in `samples/captioning/events.html`.