fix: Allow enhancement codecs to be re-tested in Capabilities.js if Settings change
Remove enhancement codecs from tested configurations so that they can be re-tested if the enhancement settings change. This PR aims to fix a minor issue that we are seeing in the Reference Player, if a stream is loaded, and afterwards the Enhancement Layer is enabled, enhancement codecs would still be filtered out due to them being in the testedCodecConfigurations. This PR removes them out from testedCodecConfigurations so they can be re-tested on next stream load.